cerowrt bql-15 smoketest support
This enables the dnsmasq dns port (due to bind9 breaking temporarily
it has been removed), and the opkg to point to the bql-15 repo.

I note that bql is a misnomer as that is merely the name of the
main kernel we are trying to track rather than actually having
bql in this release.

An early backport of bql was attempted but failed.