Make babel routing work over ipv6, but not ipv4, when nat is present
There is no easy way to detect if nat is present on an interface.
However it's possible to configure babel to at least route
ipv6 over that interface and announce ipv6 routes over it,
so you can take advantage of failover, etc.

This leads to 'interesting' paths for ipv6 vs ipv4 stuff, where
the routes can differ substantially.

Anyway, this code is mostly useful in the lab and in mesh network
scenarios, and is commented out by default.